Corrected depth calculation method The previous version of the script used the lowest 10% of PO4 headgroups as the bottom of the membrane deformation. As only a small proportion of PO4 headgroups are in the very bottom of the deformation in a Piezo1 system, this leads to systematic overestimation of the actual lowest point of the membrane.
The magnitude of the systematic error increases with increasing bilayer size,
and is expected to also vary with the amount of distortion applied to Piezo1
as a result of membrane stiffness.

I therefore redefined the bottom of the membrane deformation as simply the
lowest PO4 in the leaflet. Not using averages is justified by the fact that
the trajectory being analysed is fitted to the protein, and the bottom of the
dome is effectively the bottom of the protein, and so is a fixed, non-
fluctuating point.